feat: :sparkles: load kernel from disk
Added support for loading arbitrary files from the file system which
means we can load our kernel binary into the main memory now. The FAT
driver implementation can likely be improved and will will need to if we
want to fit more into the bootloader. Right now, we are just barely
under 512 bytes on the boot sector and we havent loaded the GDT or
switched to protected mode yet. Might have to make a second bootloader
stage or some kind of kernel pre-entry routine.

This commit also adds files related to GDB debugging a qemu session in
real mode. This takes some set up, so the config files are provided.