fix: :bug: restore stack in fat_read_file_from_fat
Forgetting to restore the stack pointer at the end of
fat_read_file_from_fat accidentally still "worked" because of the order
that registers are pushed onto the stack in the pusha instruction. The
stack was offset by 4 bytes and since the kernel address was stored
there during the previous function call, when ret was called, we
accidentally jumped straight to the kernel instead of back to our
previous function (fat_find_and_read_root_file). This achieved the
desired effect of jumping to our kernel but completely bypassed the rest
of the main function after loading the kernel from the disk.